restructuring ExtrefEditor add ControlBlockEditor interface, closes #329

move "removeAllControlBlocksAndDatasetsAndExtRefSrcBindings" from SclEditor to ControlBlockEditor, closes #329


move "analyzeDataGroups" from SclEditor to ControlBlockEditor, closes #329
